Move all MockedRemoteStorage-related out of setup
