+++ eel/tests/persistence_test.rs
@@ -13,8 +13,9 @@ mod persistence_test {
     use std::thread::sleep;
     use std::time::Duration;
 
+    use crate::setup::mocked_storage_setup::mocked_storage_node_configurable;
     use crate::setup::nigiri::NodeInstance;
-    use crate::setup::{mocked_storage_node_configurable, nigiri, NodeHandle};
+    use crate::setup::{nigiri, NodeHandle};
     use crate::try_cmd_repeatedly;
 
     const ONE_SAT: u64 = 1_000;
